--- src/main/java/com/littlewhite/lab/controller.java
@@ -1,4 +1,0 @@
-package com.littlewhite.lab;
-
-public class controller {
-}
