--- src/main/java/com/littlewhite/lab/lab1/xb/demo.java
@@ -1,19 +1,0 @@
-package com.littlewhite.lab.lab1.xb;
-
-import org.junit.Test;
-
-public class demo extends com.littlewhite.lab.lab1.CalculateSum{
-
-
-	@Override
-	public long sum(Integer... numbers){
-		return numbers.length;
-	}
-
-	@Test
-	public void myTest(){
-		demo demo = new demo();
-		super.test(demo);
-	}
-
-}
